Rename branch name warning UI component file
